Relax uv version requirement
For the current use of `uv`, it's not necessary to constrain the package version to ^0.5.0. The commands used in PyAirbyte: `uv venv` and `uv pip install` are working as well up to the current version (0.8.15) and are not likely to change up to 0.9.0.

I'm trying to use PyAirbyte with Prefect and Prefect [requires](https://github.com/PrefectHQ/prefect/blob/main/pyproject.toml#L84) 'uv>=0.6.0'. With this constraint it's impossible to have both packages installed at the same time.